Séparation runner et algorithme foobarqix
